// src/dist/sandwich_pb.d.ts
export interface IngredientMap {
  INVALID: 0;
  EGG: 1;
  HAM: 2;
  CHEESE: 3;
  LETTUCE: 4;
  TOMATO: 5;
}

export const Ingredient: IngredientMap;